Coin animation on bricks complete!
The coin animation now plays everytime the player jumps and hits any of the bricks marked as 'Coin Block' (true), which is around half of the bricks on the level at the moment butcan be easily altered and rearranged. The number of coins per brick isn't limited at the moment, and it also doesnt yet effect the players score.